fix: check for null value during form validation
